// platform/packages/ui/src/components/sfb-footer/sfb-footer.types.ts
export interface FooterNavLink {
    label: string;
    href?: string;
}

export interface ResponsibleGamblingLogo {
    src: string;
    alt: string;
    width: number;
    height: number;
    description: string;
}

export interface SfbFooterProps {
    navLinks?: FooterNavLink[];
    legalText?: string;
}